+++ R/get_tests.R
@@ -4,7 +4,7 @@
 #' Get the tests for an account. You can specify a time frame from, or to, which the tests should come (or be synced).
 #'
 #' @usage
-#' get_tests(from, to, sync = FALSE)
+#' get_tests(from, to, sync = FALSE, active = TRUE)
 #'
 #' @param from Optionally supply a time (Unix timestamp) you want the tests from. If you do not
 #' supply this value you will receive every test. This parameter is best suited for bulk exports of
@@ -18,30 +18,37 @@
 #' suited to keep your database in sync with the Hawkin database. If you do not supply this value
 #' you will receive every test.
 #'
+#' @param active There was a change to the default API configuration to reflect the majority of
+#' users API configuration. Inactive tests or tests where `active:false` are returned in these
+#' configuration. Be default, `active` is set to TRUE. To return all tests, including disabled
+#' trials, set `active` to FALSE.
+#'
 #' @return
 #' Response will be a data frame containing the trials within the time range (if specified).
 #'
 #' **id**   *str*   Test trial unique ID
 #'
+#' **active**   *logi*   The trial is active and not disabled
+#'
 #' **timestamp**   *int*   UNIX timestamp of trial
 #'
 #' **segment**   *chr*   Description of the test type and trial number of the session (testType:trialNo)
 #'
-#' **testType.id**   *chr*   Id of the test type of the trial
+#' **testType_id**   *chr*   Id of the test type of the trial
 #'
-#' **testType.name**   *chr*   Name of the test type of the trial
+#' **testType_name**   *chr*   Name of the test type of the trial
 #'
-#' **testType.canonicalId**   *chr*   Canonical Id of the test type of the trial
+#' **testType_canonicalId**   *chr*   Canonical Id of the test type of the trial
 #'
-#' **athlete.id**   *chr*   Unique Id of the athlete
+#' **athlete_id**   *chr*   Unique Id of the athlete
 #'
-#' **athlete.name**   *chr*   Athlete given name
+#' **athlete_name**   *chr*   Athlete given name
 #'
-#' **athlete.active**   *logi*   The athlete is active
+#' **athlete_active**   *logi*   The athlete is active
 #'
-#' **athlete.teams**   *list*   List containing Ids of each team the athlete is on
+#' **athlete_teams**   *list*   List containing Ids of each team the athlete is on
 #'
-#' **athlete.groups**   *list*   List containing Ids of each group the athlete is in
+#' **athlete_groups**   *list*   List containing Ids of each group the athlete is in
 #'
 #' All metrics from each test type are included as the remaining variables.
 #' If a trial does not have data for a variable it is returned NA.
@@ -67,10 +74,12 @@
 #'
 #' @importFrom rlang .data
 #' @importFrom tidyr unnest
+#' @importFrom dplyr select
+#' @importFrom dplyr relocate
 #' @export
 
 ## Get All Tests or Sync Tests -----
-get_tests <- function(from = NULL, to = NULL, sync = FALSE) {
+get_tests <- function(from = NULL, to = NULL, sync = FALSE, active = TRUE) {
 
   # Retrieve Access Token and Expiration from Environment Variables
   aToken <- base::Sys.getenv("accessToken")
@@ -99,25 +108,23 @@ get_tests <- function(from = NULL, to = NULL, sync = FALSE) {
   # From DateTime
   fromDT <- if(base::is.null(from)) {
     ""
-  } else if(!is.numeric(from)) {
-    base::stop("date must be in numeric unix format")
-  } else{
-    base::paste0("&from=",from)
+  } else if(base::is.numeric(from) && base::isTRUE(sync)) {
+    base::paste0("?syncFrom=",from)
+  } else if(base::is.numeric(from) && base::isFALSE(sync)) {
+    base::paste0("?from=",from)
   }
 
 
   # To DateTime
   toDT <- if(base::is.null(to)) {
     ""
-  } else if(!is.numeric(to)) {
-    base::stop("date must be in numeric unix format")
-  } else if(base::is.null(from)) {
+  } else if(base::is.null(from) && base::isFALSE(sync)) {
     base::paste0("?to=",to)
   } else if(base::is.null(from) && base::isTRUE(sync)) {
     base::paste0("?syncTo=",to)
-  } else if(base::isTRUE(sync)) {
+  } else if(base::is.numeric(to) && base::isTRUE(sync)) {
     base::paste0("&syncTo=",to)
-  } else {
+  } else if(base::is.numeric(to) && base::isFALSE(sync)){
     base::paste0("&to=",to)
   }
 
@@ -167,15 +174,108 @@ get_tests <- function(from = NULL, to = NULL, sync = FALSE) {
       # The code ran successfully, and 'result' contains the data frame
     }
 
+    #-----#
+    ### Create data frame ###
+    #-----#
+
     # Clean Resp Headers
     base::names(x) <- base::sub("^data\\.", "", base::names(x))
 
-    # UnNest testType and Athlete data
-    x <- x %>% tidyr::unnest(c(.data$testType, .data$athlete), names_sep = ".")
+    ##-- External IDs --##
+
+    # Create externalId df
+    extDF <- x$athlete$external
+
+    # Prepare externalId vector
+    external <- base::c()
+
+    # Identify External Ids
+    if( base::ncol(extDF) > 0) {
+      # Loop externalId columns
+      for (i in 1:base::nrow(extDF)) {
+
+        extRow <- NA
+
+        for (n in 1:base::ncol(extDF)) {
+
+          # get externalId name
+          extN <- base::names(extDF)[n]
+
+          # get ext id
+          extId <- extDF[i,n]
+
+          # create new external id name:id string
+          newExt <- base::paste0(extN, ":", extId)
+
+          # add new externalId string to row list if needed
+          extRow <- if( base::is.na(extId) ) {
+            # if extId NA, no change
+            extRow
+          } else {
+            # Add new string to extId Row
+            extRow <- if( base::is.na(extRow) ) {
+              base::paste0(newExt)
+            } else{
+              base::paste0(extRow, ",", newExt)
+            }
+          }
+
+        }
+
+        external <- base::c(external, extRow)
+      }
+    } else {
+      external <- base::rep('NA', nrow(x))
+    }
+
+    # Athlete df from original df
+    a <- x$athlete
+
+    # Remove old external from athlete df
+    a <- dplyr::select(.data = a, -dplyr::starts_with('external'))
+
+    # Bind external column to athlete df
+    a <- base::cbind(a, external)
+
+    # append Athlete prefix
+    base::names(a) <- base::paste0('athlete_', base::names(a))
+
+    ##-- Test Types --##
+
+    # Create testType df
+    t <- x$testType
+
+    # append testType prefix
+    base::names(t) <- base::paste0('testType_', base::names(t))
+
+    ##-- finish data frame --##
+
+    # select trial metadata metrics from DF
+    x1 <- dplyr::select(.data = x, base::c('id', 'timestamp', 'segment'))
+
+    # select all metrics from DF
+    x2 <- dplyr::select(.data = x, -base::c('id', 'timestamp', 'segment', 'testType', 'athlete'))
+
+    # create new complete DF
+    x <- base::cbind(x1, t, a, x2)
 
     # Clean colnames with janitor
     x <- janitor::clean_names(x)
 
+    # filter inactive tests
+    x <- if( base::isTRUE(active) ) {
+      filt <- dplyr::filter(.data = x, active == TRUE)
+
+      filt <- dplyr::relocate(.data = filt, 'active', .before = 'timestamp')
+
+      filt
+    } else if( base::isFALSE(active) ){
+      x <- dplyr::relocate(.data = x, 'active', .before = 'timestamp')
+
+      x
+    }
+
+    # return final DF
     x
   }