+++ R/get_tests_ath.R
@@ -4,7 +4,7 @@
 #' Get only tests of the specified athlete for an account.
 #'
 #' @usage
-#' get_tests_ath(athleteId, from, to)
+#' get_tests_ath(athleteId, from, to, sync = FALSE, active = TRUE)
 #'
 #' @param athleteId Supply an athlete’s id to receive tests for a specific athlete
 #'
@@ -16,10 +16,22 @@
 #' supply this value you will receive every test from the beginning of time or the optionally
 #' supplied `from` parameter. This parameter is best suited for bulk exports of historical data.
 #'
+#' @param sync  The result set will include updated and newly created tests. This parameter is best
+#' suited to keep your database in sync with the Hawkin database. If you do not supply this value
+#' you will receive every test.
+#'
+#' @param active There was a change to the default API configuration to reflect the majority of
+#' users API configuration. Inactive tests or tests where `active:false` are returned in these
+#' configuration. Be default, `active` is set to TRUE. To return all tests, including disabled
+#' trials, set `active` to FALSE.
+#'
 #' @return
-#' Response will be a data frame containing the trials from the specified team and within the time range (if specified).
+#' Response will be a data frame containing the trials from the specified team and within the time
+#' range (if specified).
 #'
 #' **id**   *str*   Test trial unique ID
+#'
+#' **active**   *logi*   The trial is active and not disabled
 #'
 #' **timestamp**   *int*   UNIX timestamp of trial
 #'
@@ -64,10 +76,12 @@
 #'
 #' @importFrom rlang .data
 #' @importFrom tidyr unnest
+#' @importFrom dplyr select
+#' @importFrom dplyr relocate
 #' @export
 
 ## Get Tests Data by Athlete Id -----
-get_tests_ath <- function(athleteId, from = NULL, to = NULL) {
+get_tests_ath <- function(athleteId, from = NULL, to = NULL, sync = FALSE, active = TRUE) {
 
   # Retrieve Access Token and Expiration from Environment Variables
   aToken <- base::Sys.getenv("accessToken")
@@ -98,7 +112,9 @@ get_tests_ath <- function(athleteId, from = NULL, to = NULL) {
     ""
   } else if(!is.numeric(from)) {
     base::stop("date must be in numeric unix format")
-  } else{
+  } else if(base::is.numeric(from) && base::isTRUE(sync)) {
+    base::paste0("&syncFrom=",from)
+  } else if(base::is.numeric(from) && base::isFALSE(sync)) {
     base::paste0("&from=",from)
   }
 
@@ -107,9 +123,12 @@ get_tests_ath <- function(athleteId, from = NULL, to = NULL) {
     ""
   } else if(!is.numeric(to)) {
     base::stop("date must be in numeric unix format")
-  } else{
+  } else if(base::is.numeric(to) && base::isTRUE(sync)) {
+    base::paste0("&syncTo=",to)
+  } else if(base::is.numeric(to) && base::isFALSE(sync)){
     base::paste0("&to=",to)
   }
+
 
   # Athlete Id
   aId <- if(base::is.character(athleteId)) {
@@ -153,18 +172,111 @@ get_tests_ath <- function(athleteId, from = NULL, to = NULL) {
     # Evaluate Response
     x <- if(resp$count[1] > 0) {
       # Convert to data frame
-      df <- base::data.frame(resp)
+      x <- base::data.frame(resp)
+
+      #-----#
+      ### Create data frame ###
+      #-----#
 
       # Clean Resp Headers
-      base::names(df) <- base::sub("^data\\.", "", base::names(df))
+      base::names(x) <- base::sub("^data\\.", "", base::names(x))
 
-      # UnNest testType and Athlete data
-      df <- df %>% tidyr::unnest(c(.data$testType, .data$athlete), names_sep = ".")
+      ##-- External IDs --##
 
-      # Clean column names with janitor
-      df <- janitor::clean_names(df)
+      # Create externalId df
+      extDF <- x$athlete$external
 
-      df
+      # Prepare externalId vector
+      external <- base::c()
+
+      # Identify External Ids
+      if( base::ncol(extDF) > 0) {
+        # Loop externalId columns
+        for (i in 1:base::nrow(extDF)) {
+
+          extRow <- NA
+
+          for (n in 1:base::ncol(extDF)) {
+
+            # get externalId name
+            extN <- base::names(extDF)[n]
+
+            # get ext id
+            extId <- extDF[i,n]
+
+            # create new external id name:id string
+            newExt <- base::paste0(extN, ":", extId)
+
+            # add new externalId string to row list if needed
+            extRow <- if( base::is.na(extId) ) {
+              # if extId NA, no change
+              extRow
+            } else {
+              # Add new string to extId Row
+              extRow <- if( base::is.na(extRow) ) {
+                base::paste0(newExt)
+              } else{
+                base::paste0(extRow, ",", newExt)
+              }
+            }
+
+          }
+
+          external <- base::c(external, extRow)
+        }
+      } else {
+        external <- base::rep('NA', nrow(x))
+      }
+
+      # Athlete df from original df
+      a <- x$athlete
+
+      # Remove old external from athlete df
+      a <- dplyr::select(.data = a, -dplyr::starts_with('external'))
+
+      # Bind external column to athlete df
+      a <- base::cbind(a, external)
+
+      # append Athlete prefix
+      base::names(a) <- base::paste0('athlete_', base::names(a))
+
+      ##-- Test Types --##
+
+      # Create testType df
+      t <- x$testType
+
+      # append testType prefix
+      base::names(t) <- base::paste0('testType_', base::names(t))
+
+      ##-- finish data frame --##
+
+      # select trial metadata metrics from DF
+      x1 <- dplyr::select(.data = x, base::c('id', 'timestamp', 'segment'))
+
+      # select all metrics from DF
+      x2 <- dplyr::select(.data = x, -base::c('id', 'timestamp', 'segment', 'testType', 'athlete'))
+
+      # create new complete DF
+      x <- base::cbind(x1, t, a, x2)
+
+      # Clean colnames with janitor
+      x <- janitor::clean_names(x)
+
+      # filter inactive tests
+      x <- if( base::isTRUE(active) ) {
+        filt <- dplyr::filter(.data = x, active == TRUE)
+
+        filt <- dplyr::relocate(.data = filt, 'active', .before = 'timestamp')
+
+        filt
+      } else if( base::isFALSE(active) ){
+        x <- dplyr::relocate(.data = x, 'active', .before = 'timestamp')
+
+        x
+      }
+
+      # return final DF
+      x
     } else {
       base::stop("No trials returned. Check athleteId or from/to entries")
     }
@@ -177,4 +289,3 @@ get_tests_ath <- function(athleteId, from = NULL, to = NULL) {
   return(Resp)
 
 }
-