+++ R/get_tests_team.R
@@ -4,7 +4,7 @@
 #' Get only tests of the specified team for an account.
 #'
 #' @usage
-#' get_tests_team(teamId, from, to)
+#' get_tests_team(teamId, from, to, sync = FALSE, active = TRUE)
 #'
 #' @param teamId Supply a team’s or a string of a comma separated list of group id’s to receive tests from
 #' specific groups. Recommended to use method `paste0()`. A maximum of 10 groups can be fetched at once.
@@ -17,10 +17,22 @@
 #' supply this value you will receive every test from the beginning of time or the optionally
 #' supplied `from` parameter. This parameter is best suited for bulk exports of historical data.
 #'
+#' @param sync  The result set will include updated and newly created tests. This parameter is best
+#' suited to keep your database in sync with the Hawkin database. If you do not supply this value
+#' you will receive every test.
+#'
+#' @param active There was a change to the default API configuration to reflect the majority of
+#' users API configuration. Inactive tests or tests where `active:false` are returned in these
+#' configuration. Be default, `active` is set to TRUE. To return all tests, including disabled
+#' trials, set `active` to FALSE.
+#'
 #' @return
-#' Response will be a data frame containing the trials from the specified team and within the time range (if specified).
+#' Response will be a data frame containing the trials from the specified team and within the time
+#' range (if specified).
 #'
 #' **id**   *str*   Test trial unique ID
+#'
+#' **active**   *logi*   The trial is active and not disabled
 #'
 #' **timestamp**   *int*   UNIX timestamp of trial
 #'
@@ -65,10 +77,12 @@
 #'
 #' @importFrom rlang .data
 #' @importFrom tidyr unnest
+#' @importFrom dplyr select
+#' @importFrom dplyr relocate
 #' @export
 
 ## Get Tests Data by Team Id -----
-get_tests_team <- function(teamId, from = NULL, to = NULL) {
+get_tests_team <- function(teamId, from = NULL, to = NULL, sync = FALSE, active = TRUE) {
 
   # Retrieve Access Token and Expiration from Environment Variables
   aToken <- base::Sys.getenv("accessToken")
@@ -97,14 +111,22 @@ get_tests_team <- function(teamId, from = NULL, to = NULL) {
   # From DateTime
   fromDT <- if(base::is.null(from)) {
     ""
-  } else {
+  } else if(!is.numeric(from)) {
+    base::stop("date must be in numeric unix format")
+  } else if(base::is.numeric(from) && base::isTRUE(sync)) {
+    base::paste0("&syncFrom=",from)
+  } else if(base::is.numeric(from) && base::isFALSE(sync)) {
     base::paste0("&from=",from)
   }
 
   # To DateTime
   toDT <- if(base::is.null(to)) {
     ""
-  } else {
+  } else if(!is.numeric(to)) {
+    base::stop("date must be in numeric unix format")
+  } else if(base::is.numeric(to) && base::isTRUE(sync)) {
+    base::paste0("&syncTo=",to)
+  } else if(base::is.numeric(to) && base::isFALSE(sync)){
     base::paste0("&to=",to)
   }
 
@@ -161,28 +183,132 @@ get_tests_team <- function(teamId, from = NULL, to = NULL) {
       # The code ran successfully, and 'result' contains the data frame
     }
 
+    #-----#
+    ### Create data frame ###
+    #-----#
+
+    ##--- Filter cases ---###
+
     # Clean Resp Headers
     base::names(x) <- base::sub("^data\\.", "", base::names(x))
-
-    # UnNest testType and Athlete data
-    x <- x %>% tidyr::unnest(c(.data$testType, .data$athlete), names_sep = ".")
 
     # Split the ID string into individual IDs
     teamIds <- base::unlist(base::strsplit(teamId, ","))
 
-    # Check if any of the IDs in teamIds are present in any of the lists in the 'athlete.teams' column
+    # Check if any of the IDs in teamIds are present in any of the lists in the 'teams' column
     filtered_df <- x %>%
-      dplyr::filter(base::any(base::sapply(.data$athlete.teams, function(ids) base::any(ids %in% teamIds))))
+      dplyr::filter(base::any(base::sapply(.data$athlete$teams, function(ids) base::any(ids %in% teamIds))))
 
-    # Clean column names with janitor
-    filtered_df <- janitor::clean_names(filtered_df)
 
+    ###
     # Use an if statement to handle the cases
     x <- if (base::nrow(filtered_df) > 0) {
       # Data matching the ID(s) was found
-      filtered_df
+      x <- filtered_df
+
+      #-----#
+      ### Create data frame ###
+      #-----#
+
+      ##-- External IDs --##
+
+      # Create externalId df
+      extDF <- x$athlete$external
+
+      # Prepare externalId vector
+      external <- base::c()
+
+      # Identify External Ids
+      if( base::ncol(extDF) > 0) {
+        # Loop externalId columns
+        for (i in 1:base::nrow(extDF)) {
+
+          extRow <- NA
+
+          for (n in 1:base::ncol(extDF)) {
+
+            # get externalId name
+            extN <- base::names(extDF)[n]
+
+            # get ext id
+            extId <- extDF[i,n]
+
+            # create new external id name:id string
+            newExt <- base::paste0(extN, ":", extId)
+
+            # add new externalId string to row list if needed
+            extRow <- if( base::is.na(extId) ) {
+              # if extId NA, no change
+              extRow
+            } else {
+              # Add new string to extId Row
+              extRow <- if( base::is.na(extRow) ) {
+                base::paste0(newExt)
+              } else{
+                base::paste0(extRow, ",", newExt)
+              }
+            }
+
+          }
+
+          external <- base::c(external, extRow)
+        }
+      } else {
+        external <- base::rep('NA', nrow(x))
+      }
+
+      # Athlete df from original df
+      a <- x$athlete
+
+      # Remove old external from athlete df
+      a <- dplyr::select(.data = a, -dplyr::starts_with('external'))
+
+      # Bind external column to athlete df
+      a <- base::cbind(a, external)
+
+      # append Athlete prefix
+      base::names(a) <- base::paste0('athlete_', base::names(a))
+
+      ##-- Test Types --##
+
+      # Create testType df
+      t <- x$testType
+
+      # append testType prefix
+      base::names(t) <- base::paste0('testType_', base::names(t))
+
+      ##-- finish data frame --##
+
+      # select trial metadata metrics from DF
+      x1 <- dplyr::select(.data = x, base::c('id', 'timestamp', 'segment'))
+
+      # select all metrics from DF
+      x2 <- dplyr::select(.data = x, -base::c('id', 'timestamp', 'segment', 'testType', 'athlete'))
+
+      # create new complete DF
+      x <- base::cbind(x1, t, a, x2)
+
+      # Clean colnames with janitor
+      x <- janitor::clean_names(x)
+
+      # filter inactive tests
+      x <- if( base::isTRUE(active) ) {
+        filt <- dplyr::filter(.data = x, active == TRUE)
+
+        filt <- dplyr::relocate(.data = filt, 'active', .before = 'timestamp')
+
+        filt
+      } else if( base::isFALSE(active) ){
+        x <- dplyr::relocate(.data = x, 'active', .before = 'timestamp')
+
+        x
+      }
+
+      # return final DF
+      x
+
     } else {
-      base::stop("No data returned. Check teamId")
+      base::stop("No data returned. Check groupId")
     }
 
     x
@@ -190,8 +316,8 @@ get_tests_team <- function(teamId, from = NULL, to = NULL) {
 
   #-----#
 
-
   # Return Response
   return(Resp)
 
 }
+