+++ NEWS.md
@@ -1,5 +1,24 @@
 # Change Log
 
+## hawkinR v1.0.3.2
+
+# bug fixes to all `get_tests_...` functions to handle NULL externalIds
+
+## hawkinR v1.0.3.1
+
+# bug fixes to get_atheletes function to handle NULL externalIds
+
+## hawkinR v1.0.3
+
+# Updates to all `get_tests_...` functions to include filter parameter for 'active' tests. Ability 
+to include or exclude disabled trials.
+
+# Updates to all `get_tests_...` functions to include 'sync' parameter. Ability 
+to utilize 'syncFrom' and 'syncTo' in all queries. This allows for more efficient queries and 
+faster download times when used frequently or to maintain an external database.
+
+# Bug fixes to externalId field
+
 ## hawkinR v1.0.2
 
 * updated `get_athlete output` to include externalId